requirements.txt

openai
requests
beautifulsoup4
pandas
scikit-learn
sentence-transformers
faiss-cpu
pyyaml
